Moved Bake function to BakeUtility
