+++ Editor/Baking.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a3d59a472f1d64131ab563d3871412ef
+guid: 4e3a2cc65be3c4b48a80696a72b3b6db
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
