+++ Editor/Baking/BakeUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b673b18abb29423dba6cd274acbd0f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
