+++ Editor/Baking/BlendShapesCurveBinding.cs.meta
