+++ Editor/Baking/BoolCurveBinding.cs
