+++ Editor/Baking/BoolCurveBinding.cs.meta
