+++ Editor/Baking/ClipBaker.cs
