+++ Editor/Baking/FloatCurveBinding.cs
