+++ Editor/Baking/FloatCurveBinding.cs.meta
