+++ Editor/Baking/QuaternionCurveBinding.cs
