+++ Editor/Baking/QuaternionCurveBinding.cs.meta
