+++ Editor/Baking/Vector3CurveBinding.cs
